Created DepartmentItem.jsx
Created the DepartmentItem component that is displayed on the ListDepartments page. Also added placeholder delete button that is to be implemented in a separate issue.

Also changed the data for accounts such that they have usernames as well as emails. These usernames will just be used to hold and display the names of the departments the accounts and emails are registered to.